requirements.txt
asgiref==3.2.10
boto3==1.17.105
botocore==1.20.105
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
decorator==5.0.9
Django==3.1.2
django-bootstrap3-datetimepicker==2.2.3
django-phonenumber-field==5.0.0
djangorestframework==3.12.4
dynaconf==3.1.4
hvac==0.10.14
idna==2.10
importlib-metadata==4.6.1
Jinja2==3.0.1
jmespath==0.10.0
knox==0.1.14
loguru==0.5.3
MarkupSafe==2.0.1
phonenumbers==8.12.12
Pillow==8.0.1
psycopg2-binary==2.8.6
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
requests==2.25.1
s3transfer==0.4.2
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
validators==0.18.2
zipp==3.5.0
